Move src files into another folder
