+++ src/math.js
